Statusbar: Gracefully adapt 3Minit Battery hook for AOSP
3Minit Battery is a free or paid third party statusbar
battery icon solution. This commit builds upon the
initial SystemUI hook source release by Gary Harrington.

We are able to significantly optimize the original code
because we are not constrained by the "one-size-fits-all"
conditions the original code must meet. At the same time,
we fully maintain the integrity of the UX intended by
the original code.

Of note is the introduction of a controller class. This
class takes over settings and state management as well
as handles receiver events. The controller class also
monitors package changes and updates feature visibility
accordingly if a needed resource is added or removed.

Upon initialization, if no feature resources are
detected, all resources are freed.

Add needed classes from DU

Signed-off-by: Martinusbe <martinusbe@gmail.com>

Conflicts:
	packages/SystemUI/src/com/android/systemui/statusbar/phone/PhoneStatusBar.java
	packages/SystemUI/src/com/android/systemui/statusbar/phone/PhoneStatusBarTransitions.java

Change-Id: I737af0f0c8f7f49abd63236128442645adb8dbf2

Conflicts:
	packages/SystemUI/src/com/android/systemui/statusbar/phone/PhoneStatusBar.java
	packages/SystemUI/src/com/android/systemui/statusbar/phone/PhoneStatusBarTransitions.java
	packages/SystemUI/src/com/android/systemui/statusbar/phone/StatusBarHeaderView.java

Conflicts:
	packages/SystemUI/src/com/android/systemui/statusbar/phone/PhoneStatusBar.java
	packages/SystemUI/src/com/android/systemui/statusbar/phone/PhoneStatusBarTransitions.java
	packages/SystemUI/src/com/android/systemui/statusbar/phone/StatusBarHeaderView.java