SystemUI: Fix navbar drawing issues in landscape
Originally StatusBarWindowView was responsible for laying out the scrims
and making sure they are full screen (behind navbar), while the other
views were respecting the navbar insets.

This was broken for two reasons for us:

 1) our scrims are more complex and aren't always direct subchildren of
    the StatusBarWindowView (mainly to display the visualizer correctly)

 2) we declared fitSystemWindows=false, so the logic to add insets was
    never actually executing.

We now split out the logic for insets into its own class and apply it to
the scrims properly.

Ref: CYNGNOS-1502
Change-Id: I5b8295d7a64b72dc911fce6d529936592f24e98b
Signed-off-by: Roman Birg <roman@cyngn.com>

Conflicts:
	packages/SystemUI/src/com/android/systemui/statusbar/phone/StatusBarWindowView.java